Move gcc_message_handlert to goto-cc folder
The only use is within goto-cc, there is no need to define this in
util/.
