+++ src/goto-cc/Makefile
@@ -7,6 +7,7 @@ SRC = armcc_cmdline.cpp \
       compile.cpp \
       cw_mode.cpp \
       gcc_cmdline.cpp \
+      gcc_message_handler.cpp \
       gcc_mode.cpp \
       gcc_version.cpp \
       goto_cc_cmdline.cpp \
