+++ src/goto-cc/as_mode.h
@@ -14,8 +14,7 @@ Date: July 2016
 #ifndef CPROVER_GOTO_CC_AS_MODE_H
 #define CPROVER_GOTO_CC_AS_MODE_H
 
-#include <util/cout_message.h>
-
+#include "gcc_message_handler.h"
 #include "goto_cc_mode.h"
 
 class as_modet:public goto_cc_modet
