+++ src/goto-cc/ld_mode.h
@@ -15,9 +15,8 @@ Date: June 2006
 #define CPROVER_GOTO_CC_LD_MODE_H
 
 #include "compile.h"
+#include "gcc_message_handler.h"
 #include "goto_cc_mode.h"
-
-#include <util/cout_message.h>
 
 #include <set>
